+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.28.0...HEAD
 
+### Changed
+
+* Make sure region and project can be passed to monitoring probes
+
 ## [0.28.0][] - 2024-05-24
 
 [0.28.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.27.0...0.28.0